Make the following changes:
,

Find this element:

  <RelationshipMapPlaceholder centralContactId="3" viewBox="-45 -73 1200 1320" scale=1 currentViewBoxOrigin={"x":-45,"y":-73}>

with the text content:

    GirlfriendMotherFatherAbhas OliFriendStudent (Computer Science)View ProfileKrishaPartnerN/AView ProfileApsara UpretyFamilyN/AView ProfileBishwa Nath OliFamilyRetired Forestry MinisterView Profile

Relevant files:

    - https://6000-firebase-studio-1746896631585.cluster-joak5ukfbnbyqspg4tewa33d24.cloudworkstations.dev/_next/static/chunks/node_modules_2c0523cf._.js
    - https://6000-firebase-studio-1746896631585.cluster-joak5ukfbnbyqspg4tewa33d24.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js
    - https://6000-firebase-studio-1746896631585.cluster-joak5ukfbnbyqspg4tewa33d24.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

why does it say father and  mother separately? it should be parents and under that it should have thier names and instead of having a tag (family) in thiere, please change to respective title like Mother , father, sister. for this and all the other contacts as well